Adding View selected Layers Component
