+++ packages/geoview-core/src/core/components/legend-2/legend-api.ts
@@ -1,6 +1,6 @@
 import { createElement } from 'react';
 import { LegendsLayerSet, api } from '@/app';
-import { LegendItem } from './legend-items/legend-item';
+import { LegendItem } from './legend-item';
 import { TypeLegendProps, TypeLegendItemProps } from './types';
 // import { LegendItemDetails, TypeLegendItemDetailsProps } from './legend-item-details/legend-item-details';
 // import { TypeLegendItemProps } from './types';
